Fix URLs for site testing on or off the domain
In ensuring that all links in the full site would work whether hosted in a subdirectory on github (githubUser.github.io/RepoName) or on a custom domain (OurSite.com) it became necessary to edit all hard-coded URLs!

There is also an environment.yml file to specify whether this is a development build or a production built.